Reworked build process to use package manager
The build process now uses a vcpkg for dependency management.

This simplifies a lot of things but required some changes to file paths.

I also redid the CMakeLists.txt and build.yml files to work with this
new build process. I added a few things to the build.yml so hopefully it
works.

Updated the documentation to reflect these changes.

I also removed some old audio files that are no longer needed.

I plan to add tests that can be ran in Git Actions but this is fine
for now.