Simplified set up and added version documentation
I removed the global download for dependencies step from the README.md
because it was unnecessary. Running the build script handled that

I also made it so the version information is now documented in
DEPENDENCIES.md. I also edited the vcpkg.json and CMakeLists.txt to
work with those versions.

I also changed the workflow file to be more accurately named

name: Windows Build and Test to -> name: Multi-Platform Build and Test