Created ResourceManager class, Improved efficiency
Use ResourceManager to load and manage fonts and textures. This removes
the need for constant calling of getRenderer from RenderWindow.hpp

It also reduces code for fixing memory leak. Rather than manually
freeing the memory for each texture, the ResourceManager class handles
it automatically when the program exits.

ResourceManager also uses an unordered_map to cache loaded textures and
fonts, improving efficiency by avoiding redundant loading operations.

The file size of Meowstro.cpp has decreased a bit thanks to this but
there are still many more improvements needed to dismantle the code
and make it more modular, maintainable, and testable.