Added GameStateManager to handle state management
This makes it easier to handle moving things around and allows me to
add additional screens/states more easily. One critical state that is
planned is the pause menu state. I will also consider a settings state.

Currently it only calls the functions in meowstro.cpp but I plan to
change that so the meowstro.cpp file isn't cluttered.