Moved game loop logic out of meowstro.cpp file
The gameLoop function continues to exist simply for reference but once
the game loop is completely replaced and moved, the 200 line function
will be deleted. It is currently unused thanks to the changes made.

There may be a small issue related to the hook animation when thrown
but that is a visual bug I will fix in the future. This code is a
working game that is essentially a rough draft.